backend/app/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
httpx==0.27.2
python-multipart==0.0.9
firebase-admin==6.5.0
requests==2.32.3


